requirements.txt
flask
psycopg2-binary
sqlalchemy
celery
redis
python-dotenv
signalwire
gunicorn
